refactor(folder): update imports to explicit submodules
Changed imports from 'store/zustand/folder' to use explicit submodules
('hooks' or 'store') in related files. Removed the unused
'store/zustand/folder/index.ts'. Updated UserPrefs imports to reference
'user-accounts' directly.
