+++ src/components/select/folders-selector.tsx
@@ -11,7 +11,7 @@ import { find, map } from 'lodash';
 import { TextUpperCase } from './color-select';
 import { FolderSelectorLabelFactory, Square } from './select-label-factory';
 import { FOLDERS } from '../../constants/folders';
-import { useRoot } from '../../store/zustand/folder';
+import { useRoot } from '../../store/zustand/folder/hooks';
 import type { FolderSelectorItem, OnChangeSelect } from '../../types/select';
 
 type FolderSelectorProps = {
