+++ src/helpers/folders.ts
@@ -7,7 +7,7 @@ import { getUserAccount } from '@zextras/carbonio-shell-ui';
 import { find } from 'lodash';
 
 import { FOLDERS } from '../constants/folders';
-import { useFolderStore } from '../store/zustand/folder';
+import { useFolderStore } from '../store/zustand/folder/store';
 import type { Folder, Folders } from '../types/folder';
 
 const NO_ACCOUNT_NAME = 'No account';
