--- src/store/zustand/folder/index.ts
@@ -1,8 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-
-export * from './store';
-export * from './hooks';
