+++ src/types/index.d.ts
@@ -5,6 +5,6 @@
  */
 
 export * from './modals';
-export * from './user-accounts.d.ts';
+export * from './user-accounts';
 export * from './soap';
 export * from './folder';
