+++ src/utils/use-prefs.ts
@@ -5,7 +5,7 @@
  */
 import { useUserSettings } from '@zextras/carbonio-shell-ui';
 
-import type { UserPrefs } from '../types';
+import { UserPrefs } from '../types/user-accounts';
 
 export const usePrefs = (): UserPrefs => {
 	const { prefs } = useUserSettings();
